+++ packages/lemonade/CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 0.1.0
 
-- Initial version.
+### Initial version.
+
+- Value validators: strings, numbers
+- Collection validators: lists, maps
+- Object validator
+- Nullable type check
+- "or", "and" validators. With shorthand operators "|" and "&"
+- Additional basic validators: "any", "equals"